Server: Convert Booking.Status to Enum

+++ uit.hotel/Businesses/BillBusiness.cs
@@ -93,7 +93,7 @@ namespace uit.hotel.Businesses
 
             foreach (var booking in billInDatabase.Bookings)
             {
-                if (booking.Status != (int)Booking.StatusEnum.CheckedOut)
+                if (booking.Status != Booking.StatusEnum.CheckedOut)
                     throw new Exception("Có phòng chưa Check-out, không thể thanh toán");
             }
